+++ src/zotero_cli/__init__.py
@@ -1,2 +1,2 @@
-VERSION = "2.8.3"
+VERSION = "2.8.4"
 __version__ = VERSION